backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic-settings>=2.7.0
httpx>=0.28.0
boto3>=1.36.0
Pillow>=11.0.0
numpy>=2.0.0
python-multipart>=0.0.18
